Move FBXParameterTypes.def to include dir
