+++ include/FBXLoader.h
@@ -7,6 +7,35 @@
 
 namespace renderbox {
 
+    union FBXPropertyValue {
+#define PARAM(TYPE, CTYPE) CTYPE TYPE;
+#include "FBXParameterTypes.def"
+        void *ptr; // For array & special types
+    };
+
+    struct FBXProperty {
+        char type;
+        FBXPropertyValue value;
+
+        ~FBXProperty() {
+            // Free non-primitive-typed value
+            switch (type) {
+                default:
+                if (value.ptr) free(value.ptr);
+                break;
+#define PARAM(TYPE, CTYPE) case (#TYPE)[0]:
+#include "FBXParameterTypes.def"
+                break;
+            }
+        };
+    };
+
+    struct FBXNode {
+        std::string name;
+        std::vector<std::unique_ptr<FBXProperty>> properties;
+        std::vector<std::unique_ptr<FBXNode>> subNodes;
+    };
+
     class FBXLoader {
     public:
 
@@ -19,6 +48,8 @@ namespace renderbox {
         std::shared_ptr<Object> destination;
 
         void enterFBXSource(std::istream &source);
+
+        void handleNode(FBXNode *node);
 
     };
